moved non-page elements to features +cleaned typos
