+++ src/pages/new.tsx
@@ -5,7 +5,7 @@ import { TopNav } from "~/layout";
 import { api } from "~/utils/api";
 import { withStaticAPIHelpers } from "~/utils/api/ssg";
 
-import SubscribeForm from "./susbscribe_form";
+import SubscribeForm from "../features/newsletter/subscribe_form";
 
 const DEFAULT_STORIES_INPUT = {
   from: 0,
